Handle things more gracefully.
Make better use of events to handle both save quitting and changing
games, as well as multiplayer more reliably.

Previously the list of horses would have continues to get longer as you
repeatedly changed games without closing the game. Now when you return
to the menu, we clear out the list of horses.

We now also scan for horses at the start of the day, this means that new
horses will get detected, without arbitrary checks throughout the day.

Managing horses is also now disabled if you are not the main player in
the game, as this would result in multiple people trying to make the
horses wander, making them more buggy than they currently are.